move JSON writing to new Writer and JsonWriter
